+++ patches/src/main/kotlin/app/revanced/patches/reddit/customclients/boostforreddit/fix/redgifs/Fingerprints.kt
@@ -1,0 +1,20 @@
+package app.revanced.patches.reddit.customclients.boostforreddit.fix.redgifs
+
+import app.revanced.patcher.fingerprint
+import com.android.tools.smali.dexlib2.AccessFlags
+import com.android.tools.smali.dexlib2.Opcode
+
+internal val createOkHttpClientFingerprint = fingerprint {
+    accessFlags(AccessFlags.PRIVATE)
+    opcodes(
+        Opcode.NEW_INSTANCE,
+        Opcode.INVOKE_DIRECT,
+        Opcode.NEW_INSTANCE,
+        Opcode.INVOKE_DIRECT,
+        Opcode.NEW_INSTANCE,
+        Opcode.INVOKE_DIRECT,
+        Opcode.INVOKE_VIRTUAL,
+        Opcode.MOVE_RESULT_OBJECT
+    )
+    custom { _, classDef -> classDef.sourceFile == "RedGifsAPIv2.java" }
+}